requirements.txt
amqp==5.3.1
asgiref==3.9.1
async-timeout==5.0.1
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
backports.asyncio.runner==1.2.0
billiard==4.2.1
celery==5.5.3
cffi==1.17.1
channels==4.3.1
channels_redis==4.3.0
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
constantly==23.10.4
cryptography==45.0.7
daphne==4.2.1
Django==5.2.6
django-bootstrap5==25.2
djangorestframework==3.16.1
exceptiongroup==1.3.0
fakeredis==2.31.1
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
iniconfig==2.1.0
kombu==5.5.4
msgpack==1.1.1
packaging==25.0
pluggy==1.6.0
prompt_toolkit==3.0.52
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
Pygments==2.19.2
pyOpenSSL==25.1.0
pyserial==3.5
pytest==8.4.2
pytest-asyncio==1.1.0
pytest-django==4.11.1
python-dateutil==2.9.0.post0
redis==6.4.0
service-identity==24.2.0
six==1.17.0
sqlparse==0.5.3
tomli==2.2.1
Twisted==25.5.0
txaio==25.6.1
typing_extensions==4.15.0
tzdata==2025.2
vine==5.1.0
wcwidth==0.2.13
zope.interface==7.2
# ---- Extra para página de controle via Jupyter/Voila ----
jupyterlab==4.2.5
notebook==7.2.2
voila==0.5.8
ipywidgets==8.1.5